// src/shared/interfaces/property/property.interfaces.ts
export type BasicSearchFiltersDef = {
  relationType: RELATION_TYPE;
  propertyType: PROPERTY_TYPE;
};

export type QueryFiltersParams = {
  rt: RELATION_TYPE;
  pt: PROPERTY_TYPE;
  pc?: PROPERTY_CONDITIONS_TYPE | null;
  city?: string;
  lat?: string | null;
  lng?: string | null;
  surface?: string | null;
  rn?: number | null;
  bn?: number | null;
  hasElevator?: boolean | null;
  pf?: PROPERTY_FURNITURE_TYPE;
  airC?: boolean | null;
  balcony?: boolean | null;
  cellar?: boolean | null;
  garage?: boolean | null;
  terrace?: boolean | null;
  wotw?: boolean | null;
  petF?: boolean | null;
  priceF?: string;
  priceT?: string;
  wria?: boolean | null;
  zoom?: string;
  pId?: string | null;
  page?: string;
};

export type QueryFiltersParamsKeys = keyof QueryFiltersParams;